Moved snapshot tests from xee_xslt_ast::instruction module to integration test suite
