+++ xee-xslt-ast/src/parse.rs
@@ -3,8 +3,9 @@ use xot::Xot;
 
 use crate::ast_core as ast;
 use crate::error::ElementError as Error;
+use crate::instruction::SequenceConstructorParser;
 use crate::staticeval::static_evaluate;
-use crate::{element::XsltParser, names::Names, state::State};
+use crate::{content::Content, context::Context, element::XsltParser, names::Names, state::State};
 
 type Result<V> = std::result::Result<V, Error>;
 
@@ -22,4 +23,23 @@ pub fn parse_transform(s: &str) -> Result<ast::Transform> {
         .map_err(|_e| Error::Unsupported)?;
     let parser = XsltParser::new(&state);
     parser.parse_transform(node)
+}
+
+pub fn parse_sequence_constructor_item(s: &str) -> Result<ast::SequenceConstructorItem> {
+    let mut xot = Xot::new();
+    let names = Names::new(&mut xot);
+    let (node, span_info) = xot.parse_with_span_info(s).unwrap();
+    let state = State::new(xot, span_info, names);
+    let node = state.xot.document_element(node).unwrap();
+
+    if let Some(element) = state.xot.element(node) {
+        let context = Context::new(state.xot.prefixes(node));
+        let content = Content::new(node, &state, context);
+        content.parse_element(
+            element,
+            ast::SequenceConstructorItem::parse_sequence_constructor_item,
+        )
+    } else {
+        Err(Error::Internal)
+    }
 }
